refactor: extract OperationSchedulerModal conflict panels + slot hooks (DEBT-1 D2-B)
Mechanical split of OperationSchedulerModal.jsx (1199 -> 728 lines) per the
DEBT-1 god-file decomposition plan. Zero behavior change; JSX/logic moved
verbatim. No test edits required.

Extracted into components/production/scheduler/:
- ConflictPanels.jsx: SlotSuggestion, ConflictAlert, PredecessorAlert,
  SuccessorConflictAlert (conflict-display subcomponents, explicit props)
- SchedulerBanners.jsx: CompatibilityWarning, ScheduleSuccess,
  UnscheduleConfirm, WizardStepHeader
- AutoPickSlotButton.jsx: PRO-gated auto-schedule subcomponent
- useSlotSuggestion.js: next-available slot fetch hook
- useResourceCompatibility.js: live resource-compatibility check hook

Modal keeps the form, state orchestration, wizard/edit-mode logic, and submit
handler. Time handling stays via toLocalInputValue.

Agent-Session: agent-d2b-schedmodal
