+++ frontend/src/components/production/OperationSchedulerModal.jsx
@@ -6,436 +6,22 @@
 import { useState, useEffect } from "react";
 import { API_URL } from "../../config/api";
 import { useResources, useResourceConflicts } from "../../hooks/useResources";
-import { formatDuration, formatTime, toLocalInputValue } from "../../utils/formatting";
+import { formatDuration, toLocalInputValue } from "../../utils/formatting";
 import Modal from "../Modal";
-
-/**
- * Shared slot suggestion pill — used by both conflict alert banners.
- */
-function SlotSuggestion({ label, nextAvailableStart, nextAvailableEnd, onUseSlot }) {
-  if (!nextAvailableStart) return null;
-  return (
-    <div className="mt-3 p-2 bg-blue-500/10 border border-blue-500/30 rounded">
-      <p className="text-sm text-blue-300">
-        {label}: {formatTime(nextAvailableStart)}
-        {nextAvailableEnd ? ` – ${formatTime(nextAvailableEnd)}` : ""}
-      </p>
-      <button
-        type="button"
-        onClick={() => onUseSlot(nextAvailableStart, nextAvailableEnd)}
-        className="mt-1 text-sm text-blue-400 hover:text-blue-300 underline"
-      >
-        Use this time
-      </button>
-    </div>
-  );
-}
-
-/**
- * Conflict alert banner with next-available slot suggestion (resource conflicts)
- */
-function ConflictAlert({
-  conflicts,
-  nextAvailableStart,
-  nextAvailableEnd,
-  onUseSlot,
-}) {
-  if (!conflicts || conflicts.length === 0) return null;
-
-  return (
-    <div className="bg-red-500/10 border border-red-500/30 rounded-lg p-4">
-      <div className="flex items-start gap-3">
-        <svg
-          className="w-5 h-5 text-red-400 flex-shrink-0 mt-0.5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-          />
-        </svg>
-        <div className="flex-1">
-          <h4 className="text-red-400 font-medium">Resource Conflict</h4>
-          <p className="text-sm text-red-400/70 mt-1">
-            This time slot overlaps with:
-          </p>
-          <ul className="mt-2 space-y-1">
-            {conflicts.map((conflict, idx) => (
-              <li key={idx} className="text-sm text-red-300">
-                - {conflict.production_order_code || conflict.po_code} -{" "}
-                {conflict.operation_code || "Operation"}
-                {conflict.scheduled_start && (
-                  <span className="text-red-400/50">
-                    {" "}
-                    ({formatTime(conflict.scheduled_start)} –{" "}
-                    {formatTime(conflict.scheduled_end)})
-                  </span>
-                )}
-              </li>
-            ))}
-          </ul>
-          <SlotSuggestion
-            label="Earliest available slot"
-            nextAvailableStart={nextAvailableStart}
-            nextAvailableEnd={nextAvailableEnd}
-            onUseSlot={onUseSlot}
-          />
-          {!nextAvailableStart && (
-            <p className="text-xs text-red-400/50 mt-2">
-              Adjust the start time or select a different resource.
-            </p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Predecessor sequence violation banner with earliest-valid-start suggestion.
- */
-function PredecessorAlert({
-  message,
-  earliestValidStart,
-  nextAvailableStart,
-  nextAvailableEnd,
-  onUseSlot,
-}) {
-  if (!message) return null;
-
-  return (
-    <div className="bg-amber-500/10 border border-amber-500/30 rounded-lg p-4">
-      <div className="flex items-start gap-3">
-        <svg
-          className="w-5 h-5 text-amber-400 flex-shrink-0 mt-0.5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-          />
-        </svg>
-        <div className="flex-1">
-          <h4 className="text-amber-400 font-medium">Sequence Constraint</h4>
-          <p className="text-sm text-amber-400/70 mt-1">{message}</p>
-          {earliestValidStart && (
-            <p className="text-sm text-amber-300 mt-1">
-              Predecessor finishes:{" "}
-              <span className="font-medium">{formatTime(earliestValidStart)}</span>
-            </p>
-          )}
-          {nextAvailableStart ? (
-            <SlotSuggestion
-              label="Earliest valid start"
-              nextAvailableStart={nextAvailableStart}
-              nextAvailableEnd={nextAvailableEnd}
-              onUseSlot={onUseSlot}
-            />
-          ) : (
-            <p className="text-xs text-amber-400/50 mt-2">
-              Schedule the predecessor operation first, then retry.
-            </p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Compatibility warning banner
- */
-function CompatibilityWarning({ reason }) {
-  if (!reason) return null;
-
-  return (
-    <div className="bg-yellow-500/10 border border-yellow-500/30 rounded-lg p-3">
-      <div className="flex items-start gap-2">
-        <svg
-          className="w-5 h-5 text-yellow-400 flex-shrink-0 mt-0.5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-          />
-        </svg>
-        <div>
-          <h4 className="text-yellow-400 font-medium text-sm">
-            Incompatible Resource
-          </h4>
-          <p className="text-sm text-yellow-400/70 mt-1">{reason}</p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Success banner shown after scheduling, with option to advance to next op
- */
-function ScheduleSuccess({ operationCode, nextOperation, onNext }) {
-  return (
-    <div className="bg-green-500/10 border border-green-500/30 rounded-lg p-4">
-      <div className="flex items-start gap-3">
-        <svg
-          className="w-5 h-5 text-green-400 flex-shrink-0 mt-0.5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M5 13l4 4L19 7"
-          />
-        </svg>
-        <div className="flex-1">
-          <h4 className="text-green-400 font-medium">
-            {operationCode} scheduled
-          </h4>
-          {nextOperation ? (
-            <div className="mt-2 flex items-center gap-3">
-              <p className="text-sm text-gray-400">
-                Next: {nextOperation.sequence} — {nextOperation.operation_code}
-              </p>
-              <button
-                type="button"
-                onClick={onNext}
-                className="text-sm text-blue-400 hover:text-blue-300 underline"
-              >
-                Schedule it now
-              </button>
-            </div>
-          ) : (
-            <p className="text-sm text-gray-400 mt-1">
-              All operations scheduled.
-            </p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Unschedule confirmation inline panel.
- */
-function UnscheduleConfirm({ onConfirm, onCancel, submitting }) {
-  return (
-    <div className="bg-amber-500/10 border border-amber-500/30 rounded-lg p-4">
-      <p className="text-amber-300 font-medium text-sm">
-        Unschedule this operation?
-      </p>
-      <p className="text-amber-400/70 text-sm mt-1">
-        The operation will return to pending and its time slot will be freed.
-      </p>
-      <div className="flex gap-3 mt-3">
-        <button
-          type="button"
-          onClick={onConfirm}
-          disabled={submitting}
-          className="px-4 py-1.5 bg-amber-600 text-white text-sm rounded-lg hover:bg-amber-700 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
-        >
-          {submitting ? "Unscheduling..." : "Confirm Unschedule"}
-        </button>
-        <button
-          type="button"
-          onClick={onCancel}
-          className="px-4 py-1.5 text-gray-400 hover:text-white text-sm transition-colors"
-        >
-          Cancel
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Successor conflict banner — shown when moving an op would violate a
- * downstream op's existing schedule.
- */
-function SuccessorConflictAlert({ successorConflicts }) {
-  if (!successorConflicts || successorConflicts.length === 0) return null;
-  return (
-    <div className="bg-orange-500/10 border border-orange-500/30 rounded-lg p-4">
-      <div className="flex items-start gap-3">
-        <svg
-          className="w-5 h-5 text-orange-400 flex-shrink-0 mt-0.5"
-          fill="none"
-          stroke="currentColor"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-          />
-        </svg>
-        <div className="flex-1">
-          <h4 className="text-orange-400 font-medium">Successor Conflict</h4>
-          <p className="text-sm text-orange-400/70 mt-1">
-            Moving this operation would violate existing schedules of:
-          </p>
-          <ul className="mt-2 space-y-1">
-            {successorConflicts.map((sc, idx) => (
-              <li key={idx} className="text-sm text-orange-300">
-                - Seq {sc.sequence} ({sc.operation_code || "Operation"})
-                {sc.scheduled_start && (
-                  <span className="text-orange-400/50">
-                    {" "}currently at {formatTime(sc.scheduled_start)}
-                  </span>
-                )}
-                {sc.earliest_valid_start && (
-                  <span className="text-orange-300">
-                    {" "}— earliest valid start: {formatTime(sc.earliest_valid_start)}
-                  </span>
-                )}
-              </li>
-            ))}
-          </ul>
-          <p className="text-xs text-orange-400/50 mt-2">
-            Reschedule the successor operations first, or choose an earlier end time.
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Auto-pick slot — PRO-gated affordance.
- *
- * Calls POST /api/v1/pro/auto-schedule with the operation + production order
- * context.  If the endpoint returns 404/403 (Core install without PRO), shows
- * a polite "PRO feature" note instead of an error.
- *
- * On success the returned slot is applied to startTime / endTime.
- */
-function AutoPickSlotButton({ operationId, productionOrderId, onSlotPicked, disabled }) {
-  const [running, setRunning] = useState(false);
-  const [proUnavailable, setProUnavailable] = useState(false);
-  const [pickError, setPickError] = useState(null);
-
-  // PRO availability is per-install, but reset on operation change so a
-  // transient 403 (e.g. session blip) doesn't stick for the whole session.
-  useEffect(() => {
-    setProUnavailable(false);
-    setPickError(null);
-  }, [operationId]);
-
-  const handleClick = async () => {
-    setRunning(true);
-    setProUnavailable(false);
-    setPickError(null);
-    try {
-      const res = await fetch(`${API_URL}/api/v1/pro/auto-schedule`, {
-        method: "POST",
-        credentials: "include",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ operation_id: operationId, production_order_id: productionOrderId }),
-      });
-      if (res.status === 404 || res.status === 403) {
-        setProUnavailable(true);
-        return;
-      }
-      if (!res.ok) {
-        let message = "Auto-pick failed — schedule manually";
-        try {
-          const errData = await res.json();
-          message = errData.detail?.message || errData.detail || message;
-        } catch {
-          // non-JSON error body — keep generic message
-        }
-        setPickError(typeof message === "string" ? message : "Auto-pick failed — schedule manually");
-        return;
-      }
-      const data = await res.json();
-      if (data.scheduled_start && data.scheduled_end) {
-        onSlotPicked(data.scheduled_start, data.scheduled_end);
-      }
-    } catch {
-      setPickError("Network error — schedule manually");
-    } finally {
-      setRunning(false);
-    }
-  };
-
-  if (proUnavailable) {
-    return (
-      <span className="text-xs text-gray-500 italic">
-        Auto-pick requires FilaOps PRO
-      </span>
-    );
-  }
-
-  return (
-    <span className="inline-flex items-center gap-2">
-      <button
-        type="button"
-        onClick={handleClick}
-        disabled={disabled || running || !operationId}
-        className="text-xs text-purple-400 hover:text-purple-300 border border-purple-500/30 hover:border-purple-400/50 rounded px-2 py-1 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-        title="Use the PRO auto-scheduler to find the optimal slot"
-      >
-        {running ? "Finding slot…" : "Auto-pick slot ✦"}
-      </button>
-      {pickError && (
-        <span className="text-xs text-amber-400">{pickError}</span>
-      )}
-    </span>
-  );
-}
-
-/**
- * Wizard step header — shown when wizardMode is true.
- */
-function WizardStepHeader({ step, total, onSkipAll }) {
-  return (
-    <div className="flex items-center justify-between px-6 py-3 bg-blue-950/30 border-b border-blue-800/30">
-      <div className="flex items-center gap-3">
-        <span className="text-xs font-medium text-blue-300 uppercase tracking-wide">
-          Schedule Wizard
-        </span>
-        <span className="text-xs text-gray-500">
-          Step {step} of {total}
-        </span>
-        {/* Progress dots */}
-        <div className="flex gap-1">
-          {Array.from({ length: total }, (_, i) => (
-            <span
-              key={i}
-              className={`w-1.5 h-1.5 rounded-full ${
-                i < step ? "bg-blue-400" : "bg-gray-700"
-              }`}
-            />
-          ))}
-        </div>
-      </div>
-      <button
-        type="button"
-        onClick={onSkipAll}
-        className="text-xs text-gray-500 hover:text-gray-300 transition-colors"
-      >
-        Skip all / later
-      </button>
-    </div>
-  );
-}
+import {
+  ConflictAlert,
+  PredecessorAlert,
+  SuccessorConflictAlert,
+} from "./scheduler/ConflictPanels";
+import {
+  CompatibilityWarning,
+  ScheduleSuccess,
+  UnscheduleConfirm,
+  WizardStepHeader,
+} from "./scheduler/SchedulerBanners";
+import AutoPickSlotButton from "./scheduler/AutoPickSlotButton";
+import { useSlotSuggestion } from "./scheduler/useSlotSuggestion";
+import { useResourceCompatibility } from "./scheduler/useResourceCompatibility";
 
 /**
  * Main modal component
@@ -575,82 +161,25 @@ export default function OperationSchedulerModal({
     }
   }, [isOpen, isEditMode, currentOp?.id]); // eslint-disable-line react-hooks/exhaustive-deps
 
-  // Check compatibility when resource changes
-  useEffect(() => {
-    setCompatWarning(null);
-    if (!resourceId || !productionOrder?.id) return;
+  // Check compatibility when resource changes (extracted hook — DEBT-1 D2-B)
+  useResourceCompatibility({
+    resourceId,
+    productionOrder,
+    resources,
+    setCompatWarning,
+  });
 
-    const selectedRes = resources.find((r) => String(r.id) === resourceId);
-    if (!selectedRes) return;
-
-    let cancelled = false;
-
-    const checkCompat = async () => {
-      try {
-        const params = new URLSearchParams({
-          resource_id: resourceId,
-          is_printer: selectedRes.is_printer ? "true" : "false",
-        });
-        const res = await fetch(
-          `${API_URL}/api/v1/production-orders/${productionOrder.id}/check-resource-compatibility?${params}`,
-          { credentials: "include" },
-        );
-        if (res.ok && !cancelled) {
-          const data = await res.json();
-          if (!data.compatible) {
-            setCompatWarning(data.reason);
-          }
-        }
-      } catch {
-        // Silently ignore - backend will still reject on submit
-      }
-    };
-
-    checkCompat();
-    return () => { cancelled = true; };
-  }, [resourceId, productionOrder?.id, resources]);
-
-  // Auto-suggest next available slot when resource is selected — skip in edit mode
-  // because the operator can see the current slot and change it deliberately.
-  useEffect(() => {
-    if (!resourceId || estimatedMinutes <= 0) return;
-    if (isEditMode) return;
-
-    const selectedRes = resources.find((r) => String(r.id) === resourceId);
-    if (!selectedRes) return;
-
-    let cancelled = false;
-
-    const fetchSuggested = async () => {
-      try {
-        const res = await fetch(
-          `${API_URL}/api/v1/production-orders/resources/next-available`,
-          {
-            method: "POST",
-            credentials: "include",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-              resource_id: parseInt(resourceId),
-              duration_minutes: estimatedMinutes,
-              is_printer: selectedRes.is_printer || false,
-            }),
-          },
-        );
-        if (res.ok && !cancelled) {
-          const data = await res.json();
-          if (data.next_available && !cancelled) {
-            setStartTime(toLocalInputValue(data.next_available));
-            setEndTime(toLocalInputValue(data.suggested_end));
-          }
-        }
-      } catch {
-        // Fall through to default time (now + duration)
-      }
-    };
-
-    fetchSuggested();
-    return () => { cancelled = true; };
-  }, [resourceId, estimatedMinutes, resources]);
+  // Auto-suggest next available slot when resource is selected — skip in edit
+  // mode because the operator can see the current slot and change it
+  // deliberately. (extracted hook — DEBT-1 D2-B)
+  useSlotSuggestion({
+    resourceId,
+    estimatedMinutes,
+    resources,
+    isEditMode,
+    setStartTime,
+    setEndTime,
+  });
 
   // Fetch operations list to find next pending op after the one just scheduled
   const fetchNextPendingOp = async (justScheduledId) => {
